+++ test/behavioral.test.js
@@ -1,18 +1,7 @@
 const waitForExpect = require('wait-for-expect');
-const { initialise } = require('../db/initialise');
-const knex = require('../db/connect')
-
-function delay(ms) {
-  console.log(`Waiting for ${ms} milliseconds...`);
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-async function queryFactory(testQuery , testFunctions) {
-  await waitForExpect(async () => {
-    const query = await knex.raw(`${testQuery}`);
-    testFunctions(query.rows);
-  }, 30000, 1000);
-}
+const { initialise } = require('../testDB/initialise');
+const knex = require('../testDB/connect')
+const { delay, queryFactory } = require('./util');
 
 describe('Test Lucid chart Importer ', () => {
   beforeAll(async (done) => {
